Extract markdown code component
