+++ app/markdown/application_markdown.rb
@@ -13,6 +13,9 @@ class ApplicationMarkdown < MarkdownRails::Renderer::Rails
   delegate \
     :link_to,
     :inline_svg,
+    :code_block_component,
+    :code_formatter,
+    :clipboard_copy,
     to: :helpers
 
   def extensions
@@ -50,32 +53,7 @@ class ApplicationMarkdown < MarkdownRails::Renderer::Rails
   end
 
   def block_code(code, metadata)
-    language, filename = metadata.split(":") if metadata
-
-    lexer = Rouge::Lexer.find(language) || Rouge::Lexers::PlainText
-
-    tag.div(class: "code-wrapper highlight language-#{language}") do
-      header = tag.div
-      if filename
-        header = tag.div(class: "code-header") do
-          html = inline_svg("app-dots.svg", class: "app-dots")
-          if filename
-            html += tag.span(filename, class: "code-filename")
-          end
-          html
-        end
-      end
-
-      body = tag.div(class: "code-body") do
-        tag.pre do
-          tag.code do
-            raw code_formatter.format(lexer.lex(code))
-          end
-        end + clipboard_copy(code)
-      end
-
-      header + body
-    end
+    code_block_component(code, metadata)
   end
 
   def image(link, title, alt_text)
@@ -89,14 +67,6 @@ class ApplicationMarkdown < MarkdownRails::Renderer::Rails
   end
 
   private
-
-  def clipboard_copy(text)
-    render partial: "components/clipboard_copy", locals: {text: text}
-  end
-
-  def code_formatter
-    @code_formatter ||= Rouge::Formatters::HTML.new
-  end
 
   # This is provided as an example; there's many more YouTube URLs that this wouldn't catch.
   def youtube_tag(url, alt)
