+++ app/views/components/code/_wasm.html.erb
@@ -1,0 +1,3 @@
+<script type="text/javascript">
+
+</script>
