--- app/views/wasm/_hello_world.html.erb
@@ -1,7 +1,0 @@
-<script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@ruby/3.3-wasm-wasi@2.5.0/dist/browser.script.iife.js"></script>
-<script type="text/ruby">
-  require "js"
-
-  puts RUBY_VERSION # => Hello, world! (printed to the browser console)
-  JS.global[:document].write "Hello, world!"
-</script>
